+++ cmd/oh/segment/sessionName/sessionName.go
@@ -1,17 +1,23 @@
 package sessionName
 
 import (
+	"net/url"
+	"path/filepath"
+
+	"crdx.org/io/cmd/oh/link"
 	"crdx.org/io/cmd/oh/segment"
 	"crdx.org/io/cmd/oh/style"
 	"crdx.org/io/session"
 )
 
 type state struct {
-	name  string
-	emoji string
+	name        string
+	emoji       string
+	address     string
+	isPersisted func() bool
 }
 
-func New(name string) segment.Factory {
+func New(name string, directory string, isPersisted func() bool) segment.Factory {
 	return func(options segment.Options) (segment.Segment, error) {
 		var args struct {
 			Emoji bool `toml:"emoji"`
@@ -26,14 +32,23 @@ func New(name string) segment.Factory {
 			emoji = session.Emoji(name)
 		}
 
-		return state{name: name, emoji: emoji}, nil
+		address := ""
+		if directory != "" {
+			address = (&url.URL{Scheme: "file", Path: filepath.ToSlash(directory)}).String()
+		}
+
+		return state{name: name, emoji: emoji, address: address, isPersisted: isPersisted}, nil
 	}
 }
 
 func (self state) Render(segment.Context) string {
-	if self.emoji == "" {
-		return style.Subtle(self.name)
+	text := style.Subtle(self.name)
+	if self.address != "" && self.isPersisted != nil && self.isPersisted() {
+		text = link.RenderURL(text, self.address)
+	}
+	if self.emoji != "" {
+		text += " " + self.emoji
 	}
 
-	return style.Subtle(self.name) + " " + self.emoji
+	return text
 }